All tests passing
need to pass by again and check testing logic/ add delete order by id tests

+++ src/main/java/com/example/service/OrderService.java
@@ -28,6 +28,10 @@ public class OrderService extends MainService {
         return orderRepository.getOrderById(orderId);
     }
     public void deleteOrderById(UUID orderId) throws IllegalArgumentException {
+        Order order = orderRepository.getOrderById(orderId);
+        if(order == null){
+            throw new RuntimeException("Order not found");
+        }
         orderRepository.deleteOrderById(orderId);
     }
 }
